+++ api/rust/cargo/lief/tests/macho_tests.rs
@@ -50,6 +50,11 @@ fn explore_macho(_: &str, macho: &lief::macho::Binary) {
         format!("{:?}", binding);
     }
 
+    for stub in macho.symbol_stubs() {
+        format!("{stub:?}");
+        format!("{}", stub.raw().len());
+    }
+
     for command in macho.commands() {
         format!("{command:?}");
         match command {