+++ include/LIEF/MachO.hpp
@@ -64,6 +64,7 @@
 #include "LIEF/MachO/SegmentCommand.hpp"
 #include "LIEF/MachO/SegmentSplitInfo.hpp"
 #include "LIEF/MachO/SourceVersion.hpp"
+#include "LIEF/MachO/Stub.hpp"
 #include "LIEF/MachO/SubClient.hpp"
 #include "LIEF/MachO/SubFramework.hpp"
 #include "LIEF/MachO/Symbol.hpp"